backend/pyproject.toml
[tool.poetry]
name = "leitor-documentos-granto"
version = "0.1.0"
description = ""
authors = ["antonio <aaba03@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
llama-index-core = "^0.10.43"
llama-index-readers-file = "^0.1.23"
llama-index-llms-ollama = "^0.1.5"
llama-index-embeddings-huggingface = "^0.2.1"
python-multipart = "^0.0.9"
python-dotenv = "^1.0.1"
llama-parse = "^0.4.4"
llama-index-postprocessor-flag-embedding-reranker = "^0.1.3"
flagembedding = {git = "https://github.com/FlagOpen/FlagEmbedding.git"}
ipykernel = "^6.29.4"
llama-index-embeddings-openai = "^0.1.10"
llama-index = "^0.10.43"
pypdf = "^4.2.0"
sqlalchemy = "^2.0.30"
psycopg2 = "^2.9.9"
pdf2image = "^1.17.0"
llama-index-vector-stores-chroma = "^0.1.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
